requirements.txt
discord.py
python-dotenv
graypy